First refactor step.
